pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "account-hub"
version = "0.1.0"
description = "CLI-based identity aggregation tool for linking email accounts and discovering associated services"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    # CLI
    "typer[all]>=0.12",
    "rich>=13.9",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # Auth / OAuth
    "authlib>=1.3",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0,<5.0",
    # Encryption
    "cryptography>=43.0",
    # HTTP
    "httpx>=0.27",
    # Config
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.21",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.13",
    "coverage>=7.6",
]

[project.scripts]
accounthub = "account_hub.cli.main:app"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.9"
strict = true
